+++ test/ciphers/feal/integration/test_feal_option_encoding.py
@@ -3,7 +3,7 @@ import unittest
 # noinspection PyUnresolvedReferences
 import test.context
 from ciphers.feal import feal
-from test.ciphers.feal.integration.wrappers import default_encrypt_args, default_decrypt_args
+from test.ciphers.feal.integration.patchers import default_encrypt_args, default_decrypt_args
 
 
 class TestFEALOptionEncoding(unittest.TestCase):
