+++ test/ciphers/feal/integration/test_feal_option_round_number.py
@@ -3,7 +3,7 @@ import unittest
 # noinspection PyUnresolvedReferences
 import test.context
 from ciphers.feal import feal, FEALArgumentException
-from test.ciphers.feal.integration.wrappers import default_decrypt_args, default_encrypt_args
+from test.ciphers.feal.integration.patchers import default_decrypt_args, default_encrypt_args
 
 
 class TestFEALOptionRoundNumber(unittest.TestCase):
